Refactor Dialogs towards reusable Components
