+++ src/index.js
@@ -139,8 +139,12 @@ export { default as BAMListingPage } from './Pages/BAMListing';
 export { default as BAMListingPageStory } from './Pages/BAMListingPageStory';
 export { default as TABLE_COLUMN_TYPES } from './Table/ColumnTypes';
 export { default as Tooltip } from './MaterialComponents/Tooltip';
-export { default as ConfirmationDialog } from './MaterialComponents/Dialog';
-export { default as PromptDialog } from './MaterialComponents/Dialog/Prompt';
+export {
+  default as ConfirmationDialog,
+} from './MaterialComponents/Dialog/Prompt';
+export {
+  default as PromptDialog,
+} from './MaterialComponents/Dialog/Prompt/Number';
 export { default as MassiveChangeDialog } from './Dialog/MassiveChange';
 export { default as Dialog } from './Dialog';
 export { default as IconVisible } from './MaterialComponents/Icons/IconVisible';
